+++ webserver/script/tests/__init__.py
@@ -1,0 +1,9 @@
+# ref:
+# (1) https://www.django-rest-framework.org/api-guide/testing/
+# (2) https://www.django-rest-framework.org/api-guide/routers/#routing-for-extra-actions
+# (3) https://docs.djangoproject.com/en/2.2/topics/db/examples/many_to_one/
+
+# Create your tests here.
+
+from script.tests.test_data import *
+from script.tests.test_statistics import *
