+++ webserver/script/tests/test_statistics.py
@@ -1,57 +1,10 @@
 from django.test import TestCase
 from rest_framework import status
 from rest_framework.reverse import reverse
-from rest_framework.test import APITestCase, APIClient
+from rest_framework.test import APITestCase
 from script.models.data import County
 from script.models.statistics import Energy
-
-# Create your tests here.
-# ref:
-# (1) https://www.django-rest-framework.org/api-guide/testing/
-# (2) https://www.django-rest-framework.org/api-guide/routers/#routing-for-extra-actions
-# (3) https://docs.djangoproject.com/en/2.2/topics/db/examples/many_to_one/
-
-
-def create_county(county_name, residents):
-    client = APIClient()
-    url = reverse('county-list')
-    data = {
-        'name': county_name,
-        'residents': residents
-    }
-    response = client.post(url, data, format='json')
-    return response
-
-
-def create_energy(county_name, year, month, energy):
-    client = APIClient()
-    url = reverse('energy-list')
-    data = {
-        'county': county_name,
-        'year': year,
-        'month': month,
-        'energy': energy
-    }
-    response = client.post(url, data, format='json')
-    return response
-
-
-class CountyTests(APITestCase):
-
-    county_name = 'Santa Cruz'
-    residents = 9812
-
-    def test_create_county(self):
-        """
-        Ensure we can create a new county object.
-        """
-        response = create_county(self.county_name, self.residents)
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(County.objects.count(), 1)
-        obj = County.objects.get()
-        self.assertEqual(obj.name, self.county_name)
-        self.assertEqual(obj.residents, self.residents)
-
+from script.tests.utils import create_county, create_energy
 
 class EnergyTests(APITestCase):
